requirements_dev.txt
-r requirements.txt
coverage
django-extensions
factory_boy
faker
ipdb
pylint
pylint-django
unidecode
